// src/core/interfaces/familyDetails.ts
export interface IFamilyDetails {
  fatherName: string;
  fatherDOB: Date;
  motherName: string;
  motherDOB: Date;
  martialStatus: string;
  spouseName?: string;
  spouseDOB?: Date;
  numberOfChildren?: number;
  child1_Name?: string;
  child1_dob?: Date;
  child2_Name?: string;
  child2_dob?: Date;
  child3_Name?: string;
  child3_dob?: Date;
  child4_Name?: string;
  child4_dob?: Date;
}